refactor: tide up code for single select and multi select components
