+++ workspaces/core/src/select/private/useSelect.ts
@@ -1,31 +1,7 @@
 import {useCallback, useEffect, useRef, useState} from "react";
+import checkNonSelectedIndex from "./checkNonSelectedIndex";
 import Option from "./types/OptionProps";
 import {MultiSelectProps, SelectProps, SingleSelectProps} from "./types/SelectProps";
-
-const checkNonSelectedIndex = (index: number, length: number, selectedIndexes: number[], direction: 'up' | 'down'): number => {
-  const nonSelectedIndexes: number[] = [];
-  let closestNonSelectedIndex = 9999;
-  let value = index;
-  // Array of selected indexes is empty or doesn't contain provided index
-  if (!selectedIndexes?.length || !selectedIndexes?.includes(index)) return value;
-
-  for (let i = length; i >= 0; i--) {
-    // We are looking for number that's not in selectedIndexes array
-    if (!selectedIndexes?.includes(i)) {
-      nonSelectedIndexes.push(i);
-      // We are looking for closest number higher than provided index
-      if (direction === 'down' && i > index && i - index < closestNonSelectedIndex - index) closestNonSelectedIndex = i;
-      // We are looking for closest number lower than provided index
-      else if (direction === 'up' && i < index && index - i < closestNonSelectedIndex - index) closestNonSelectedIndex = i;
-    }
-  }
-  if (direction === 'down' && closestNonSelectedIndex === 9999) closestNonSelectedIndex = Math.min(...nonSelectedIndexes);
-  if (direction === 'up' && closestNonSelectedIndex === 9999) closestNonSelectedIndex = Math.max(...nonSelectedIndexes);
-
-  value = closestNonSelectedIndex;
-
-  return value;
-};
 
 interface Props {
   options: SelectProps["options"],
@@ -49,7 +25,6 @@ const useSelect = ({
  onChange,
  value,
  hoveredIndex = -1,
- selectedIndex = -1,
  selectedIndexes = [],
  isSearchable,
  disabled,
